Cleanup icon decomposition leftovers
