+++ src/components/Button/index.tsx
@@ -1,5 +1,6 @@
 import ActivityIndicator from '@components/ActivityIndicator';
 import Icon from '@components/Icon';
+import type IconSize from '@components/Icon/types';
 import type {PressableRef} from '@components/Pressable/GenericPressable/types';
 import PressableWithFeedback from '@components/Pressable/PressableWithFeedback';
 import Text from '@components/Text';
@@ -235,6 +236,20 @@ function KeyboardShortcutComponent({
     return null;
 }
 
+/** Maps the button's size booleans to the matching `Icon` preset size. */
+function getIconSize({extraSmall, small, large}: {extraSmall: boolean; small: boolean; large: boolean}): IconSize {
+    if (extraSmall) {
+        return CONST.ICON_SIZE.EXTRA_SMALL;
+    }
+    if (small) {
+        return CONST.ICON_SIZE.SMALL;
+    }
+    if (large) {
+        return CONST.ICON_SIZE.LARGE;
+    }
+    return CONST.ICON_SIZE.MEDIUM;
+}
+
 function Button({
     allowBubble = false,
 
@@ -305,6 +320,8 @@ function Button({
     const styles = useThemeStyles();
     const StyleUtils = useStyleUtils();
     const [isHovered, setIsHovered] = useState(false);
+
+    const iconSize = getIconSize({extraSmall, small, large});
 
     const buttonLoadingReasonAttributes: SkeletonSpanReasonAttributes = {
         context: 'Button',
@@ -382,14 +399,7 @@ function Button({
                                 <Icon
                                     src={icon}
                                     fill={isHovered ? (iconHoverFill ?? defaultFill) : (iconFill ?? defaultFill)}
-                                    // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                    extraSmall={extraSmall}
-                                    // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                    small={small}
-                                    // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                    medium={medium}
-                                    // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                    large={large}
+                                    size={iconSize}
                                     isButtonIcon
                                     accessibilityLabel={iconAccessibilityLabel}
                                 />
@@ -402,14 +412,7 @@ function Button({
                             <Icon
                                 src={iconRight ?? icons.ArrowRight}
                                 fill={isHovered ? (iconRightHoverFill ?? iconHoverFill ?? defaultFill) : (iconRightFill ?? iconFill ?? defaultFill)}
-                                // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                extraSmall={extraSmall}
-                                // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                small={small}
-                                // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                medium={medium}
-                                // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                                large={large}
+                                size={iconSize}
                                 isButtonIcon
                             />
                         </View>
