+++ src/components/Search/SearchList/ListItem/ExpandCollapseArrowButton.tsx
@@ -35,8 +35,7 @@ function ExpandCollapseArrowButton({isExpanded, onPress}: ExpandCollapseArrowBut
                     src={isExpanded ? icons.UpArrow : icons.DownArrow}
                     fill={theme.icon}
                     additionalStyles={!hovered && styles.opacitySemiTransparent}
-                    // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                    small={isLargeScreenWidth}
+                    size={isLargeScreenWidth ? CONST.ICON_SIZE.SMALL : undefined}
                 />
             )}
         </PressableWithFeedback>
