+++ src/components/Search/SearchPageHeader/SearchDisplayDropdownButton.tsx
@@ -1,4 +1,5 @@
 import Icon from '@components/Icon';
+import type IconSize from '@components/Icon/types';
 import {PressableWithFeedback} from '@components/Pressable';
 import {ListFilterHeightContextProvider} from '@components/Search/FilterComponents/ListFilterHeightContext';
 import DisplayPopup from '@components/Search/FilterDropdowns/DisplayPopup';
@@ -53,6 +54,13 @@ function SearchDisplayDropdownButton({queryJSON, searchResults, onSort}: SearchD
     );
 
     if (shouldUseNarrowLayout || isMediumScreenWidth) {
+        let iconSize: IconSize | undefined;
+        if (isMediumScreenWidth) {
+            iconSize = CONST.ICON_SIZE.EXTRA_SMALL;
+        } else if (shouldUseNarrowLayout) {
+            iconSize = CONST.ICON_SIZE.SMALL;
+        }
+
         return (
             <FilterPopupButton
                 PopoverComponent={displayPopup}
@@ -70,10 +78,7 @@ function SearchDisplayDropdownButton({queryJSON, searchResults, onSort}: SearchD
                         <Icon
                             src={expensifyIcons.Gear}
                             fill={theme.icon}
-                            // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                            small={shouldUseNarrowLayout}
-                            // eslint-disable-next-line @typescript-eslint/no-deprecated -- legacy icon sizing
-                            extraSmall={isMediumScreenWidth}
+                            size={iconSize}
                         />
                     </PressableWithFeedback>
                 )}
